Refactor EMA strategy into modular strategy hooks
